chore: rename mcm to many-chain-multisig
