+++ chains/solana/contracts/Anchor.toml
@@ -15,7 +15,7 @@ example_ccip_sender = "4LfBQWYaU6zQZbDyYjX8pbY4qjzrhoumUFYZEZEqMNhJ"
 lockrelease_token_pool = "8eqh8wppT9c5rw4ERqNCffvU6cNFJWff9WmkcYtmGiqC"
 external_program_cpi_stub = "2zZwzyptLqwFJFEFxjPvrdhiGpH9pJ3MfrrmZX6NTKxm"
 fee_quoter = "FeeQPGkKDeRV1MgoYfMH6L8o3KeuYjwUZrgn4LRKfjHi"
-mcm = "5vNJx78mz7KVMjhuipyr9jKBKcMrKYGdjGkgE4LUmjKk"
+many_chain_multisig = "5vNJx78mz7KVMjhuipyr9jKBKcMrKYGdjGkgE4LUmjKk"
 rmn_remote = "RmnXLft1mSEwDgMKu2okYuHkiazxntFFcZFrrcXxYg7"
 test_ccip_invalid_receiver = "FmyF3oW69MSAhyPSiZ69C4RKBdCPv5vAFTScisV7Me2j"
 test_ccip_receiver = "EvhgrPhTDt4LcSPS2kfJgH6T6XWZ6wT3X9ncDGLT1vui"
