+++ chains/solana/contracts/programs/many-chain-multisig/Cargo.toml
@@ -1,12 +1,12 @@
 [package]
-name = "mcm"
+name = "many-chain-multisig"
 version = "0.1.0-dev"
 description = "SVM implementation of ManyChainMultiSig"
 edition = "2021"
 
 [lib]
 crate-type = ["cdylib", "lib"]
-name = "mcm"
+name = "many_chain_multisig"
 
 [features]
 no-entrypoint = []
