+++ chains/solana/contracts/programs/many-chain-multisig/src/instructions/set_config.rs
