+++ chains/solana/contracts/programs/many-chain-multisig/src/state/config.rs
