+++ chains/solana/gobindings/many_chain_multisig/AcceptOwnership.go
