+++ chains/solana/gobindings/many_chain_multisig/AppendSignatures.go
