+++ chains/solana/gobindings/many_chain_multisig/ClearSigners_test.go
