+++ chains/solana/gobindings/many_chain_multisig/InitSignatures.go
