+++ chains/solana/gobindings/many_chain_multisig/InitSigners.go
