+++ chains/solana/gobindings/many_chain_multisig/Initialize.go
