+++ chains/solana/gobindings/many_chain_multisig/Initialize_test.go
