+++ chains/solana/gobindings/many_chain_multisig/SetConfig.go
