+++ chains/solana/gobindings/many_chain_multisig/SetConfig_test.go
