+++ chains/solana/gobindings/many_chain_multisig/SetRoot.go
