+++ chains/solana/gobindings/many_chain_multisig/accounts.go
