+++ chains/solana/gobindings/many_chain_multisig/types.go
